requirements.txt
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.3
django-allauth==65.3.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-rest-knox==5.0.2
djangorestframework==3.15.2
drf-yasg==1.21.8
exceptiongroup==1.2.2
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.2
tomli==2.2.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
